# Тили-тили/Тили-тили_API_openapi.yaml
openapi: 3.0.3
info:
  title: Тили-тили API
  version: 0.28.0
  description: |
    Контракт бэкенда для PWA «Тили-тили» (tili-tili.ru) — свадебный планировщик.
    Фронтенд (React) уже работает на моках; этот файл — единственный источник правды для бэкенда (Claude Code).

    ## Решения владельца 2026-09-02 (контракт v0.2)
    - Деньги — объект `{ amount: копейки, currency: ISO 4217 }` у КАЖДОЙ суммы.
      Не одна валюта на свадьбу: план §19.7 описывает свадьбу за границей, где
      сделка в евро, а бюджет в рублях. В MVP принимается только RUB.
    - У слота нет собственного статуса. Слот либо пуст, либо несёт сделку;
      состояние живёт в `Deal.state` — шесть значений. `Slot.tileState` —
      производная подпись для мозаики, только для чтения.
    - Гость опознаётся ОДНИМ персональным токеном во всех гостевых путях
      (`/rsvp/{guestToken}`, `/gifts/{guestToken}`, `/join/{guestToken}/…`).
      Общий код свадьбы не годится: по нему нельзя подставить имя в приглашение
      и нельзя дать гостю снять СВОЙ резерв подарка. Анонимность сохраняется —
      система знает гостя, а API пары этот токен не отдаёт никогда (§9).
    - Переноса данных с localStorage-моков не будет: реальных пользователей не
      было, эндпоинт импорта не заводится.

    ## Роли
    - `couple` — молодожёны (владельцы свадьбы): полный доступ.
    - `helper` — помощник (друг/родственник): орг-разделы, без бюджета и сделок.
...
    - `vendor` — подрядчик: свои сделки, лиды, чаты, анкета.
...

...
servers:
...
paths:
...
  /users/me/consent:
    post:
      tags: [auth]
      summary: Зафиксировать согласие на обработку персональных данных
      description: |
        152-ФЗ: согласие даётся явным действием. Сохраняются версия документа,
        дата и IP — это и есть подтверждение. Предустановленная галочка согласием не является.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policyVersion]
              properties:
                policyVersion: { type: string, example: '2026-09-02' }
      responses:
        '201': { description: Зафиксировано }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [auth]
      summary: Отозвать согласие
      description: Равносильно удалению аккаунта и всех данных свадьбы.
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: OK }


  # ───────────────────────── ПРОФИЛЬ ПОЛЬЗОВАТЕЛЯ ─────────────────────────
  /users/me:
    get:
      tags: [auth]
      summary: Мой профиль
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
    patch:
      tags: [auth]
      summary: Изменить профиль и настройки уведомлений
      description: Имя, язык, таймзона, четыре канала push и тихие часы — экран «Настройки».
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserProfile' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserProfile' }
    delete:
      tags: [auth]
      summary: Удалить аккаунт
      description: |
        Soft-delete на 30 дней, затем полное удаление кроном (План §19.1).
        Активные сделки требуют подтверждения второй стороны — 409 со списком.
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: OK }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/me/sessions:
    get:
      tags: [auth]
      summary: Мои сессии и устройства
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Session' }
    delete:
      tags: [auth]
      summary: Выйти со всех устройств, кроме текущего
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: OK }
  /users/me/sessions/{sessionId}:
    delete:
      tags: [auth]
      summary: Завершить сессию
      description: 'Выход со всех устройств сразу — DELETE /users/me/sessions.'
      security: [{ bearerAuth: [] }]
      parameters: [{ name: sessionId, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: OK }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /users/me/export:
    get:
      tags: [auth]
      summary: Экспорт моих данных
      description: 152-ФЗ, право на получение данных. JSON со всем, что принадлежит пользователю.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: object }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Конфликт (уже существует)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
              description: |
                Только у 422: какие именно поля не прошли проверку.
                Ключ — имя поля (`reason`, `action`, `synonyms.тамада`),
                значение — что с ним не так. Человеку показывается `message`,
                поля подсвечивает форма.
...
    UserProfile:
      type: object
      properties:
        id: { type: string, readOnly: true }
        name: { type: string }
        phone: { type: string, readOnly: true }
        email:
          type: string
          nullable: true
          readOnly: true
          description: |
            Пока только читается: отправителя писем нет (хвост владельца), и
            принимать адрес, на который ничего не уйдёт, — обещать доставку.
            Правка появится вместе с отправителем и подтверждением адреса.
        isStaff:
          type: boolean
          readOnly: true
          description: |
            Сотрудник платформы: по этому признаку в меню «Мы» появляется
            «Админка». Только чтение и только про себя: списка сотрудников
            наружу нет, а признак ставится руками в базе при найме.
            PATCH /users/me берёт тело по этой же схеме — без readOnly контракт
            объявил бы права настраиваемыми, то есть «сделай меня админом»
            в один запрос.
        lang: { type: string, enum: [ru, en] }
        tz: { type: string, example: Asia/Yekaterinburg, description: 'нужна для открытия чата дня X в 09:00 по местному' }
        push:
          type: object
          description: четыре независимых канала — тумблеры в настройках
          properties:
            tasks: { type: boolean }
            chats: { type: boolean }
            deals: { type: boolean }
            tips: { type: boolean }
        quietHours:
          type: object
          description: 'по умолчанию 22:00–09:00; в день X отключаются автоматически'
          properties:
            from: { type: string, example: '22:00' }
            to: { type: string, example: '09:00' }
    Session:
      type: object
      properties:
        id: { type: string }
        device: { type: string, example: 'iPhone · Safari' }
        current: { type: boolean }
        createdAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []